refactor(runtime): extract tool-spec build into agent/turn — #7415
Move the per-iteration tool-spec assembly (registry filter + activated
deferred tools + known-name set + native-tools flag) verbatim into
turn/tool_specs.rs as build_iteration_tool_specs(..) ->
IterationToolSpecs; the loop destructures the result into the same
local names.
